+++ src/skills/skills.service.ts
@@ -1,6 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { type DeleteResult, Repository, type UpdateResult } from 'typeorm';
+import { Repository } from 'typeorm';
 
 import { type CreateSkillDto } from './dto/create-skill.dto';
 import { type UpdateSkillDto } from './dto/update-skill.dto';
@@ -14,6 +18,14 @@ export class SkillsService {
   ) {}
 
   async create(createSkillDto: CreateSkillDto): Promise<Skill> {
+    const skillExist = await this.skillRepository.findOneBy({
+      name: createSkillDto.name,
+    });
+
+    if (skillExist) {
+      throw new ConflictException('Skill already exists');
+    }
+
     const skill = this.skillRepository.create(createSkillDto);
     return await this.skillRepository.save(skill);
   }
@@ -23,17 +35,34 @@ export class SkillsService {
   }
 
   async findOne(id: string): Promise<Skill> {
-    return await this.skillRepository.findOneBy({ id });
+    const skill = await this.skillRepository.findOneBy({ id });
+
+    if (!skill) {
+      throw new NotFoundException('Skill not found');
+    }
+
+    return skill;
   }
 
-  async update(
-    id: string,
-    updateSkillDto: UpdateSkillDto,
-  ): Promise<UpdateResult> {
-    return await this.skillRepository.update(id, updateSkillDto);
+  async update(id: string, updateSkillDto: UpdateSkillDto): Promise<Skill> {
+    const skill = await this.findOne(id);
+
+    if (!skill) {
+      throw new NotFoundException('Skill not found');
+    }
+
+    await this.skillRepository.update(id, updateSkillDto);
+
+    return await this.findOne(id);
   }
 
-  async remove(id: string): Promise<DeleteResult> {
-    return await this.skillRepository.delete(id);
+  async remove(id: string): Promise<void> {
+    const skill = await this.findOne(id);
+
+    if (!skill) {
+      throw new NotFoundException('Skill not found');
+    }
+
+    await this.skillRepository.delete(id);
   }
 }
